+++ components/Builder.tsx
@@ -1,5 +1,3 @@
-import { SectionHeader } from "components/SectionHeader/SectionHeader";
-import { BaseOuterWrapper } from "components/style/Wrappers";
 import { laptopAndUnder, mobileAndUnder, tabletAndUnder } from "constant";
 import { useInView } from "framer-motion";
 import { useAddHashToUrl } from "hooks/helpers/useAddHashToUrl";
@@ -7,10 +5,12 @@ import dynamic from "next/dynamic";
 import OO from "public/assets/oo-logo.svg";
 import { useRef } from "react";
 import styled from "styled-components";
+import { SectionHeader } from "./SectionHeader";
+import { BaseOuterWrapper } from "./Wrappers";
 
-const Tabs = dynamic(() => import("components/Tabs/Tabs"));
+const Tabs = dynamic(() => import("components/Tabs"));
 
-export function Builder() {
+export default function Builder() {
   const id = "builder";
   const ref = useRef<HTMLDivElement>(null);
   const inView = useInView(ref, { amount: "some" });
