+++ components/Footer.tsx
@@ -1,5 +1,3 @@
-import { MailChimpForm, VoteTicker } from "components";
-import { BaseOuterWrapper } from "components/style/Wrappers";
 import { footerLinks, laptopAndUnder, mobileAndUnder, socialLinks, tabletAndUnder } from "constant";
 import { useAddHashToUrl } from "hooks/helpers/useAddHashToUrl";
 import NextLink from "next/link";
@@ -8,8 +6,11 @@ import UpRightArrowBlack from "public/assets/up-right-arrow-black.svg";
 import { useRef } from "react";
 import styled from "styled-components";
 import { isExternalLink } from "utils";
+import { MailChimpForm } from "./MailChimpForm";
+import { VoteTicker } from "./VoteTicker";
+import { BaseOuterWrapper } from "./Wrappers";
 
-export function Footer() {
+export default function Footer() {
   const id = "contact";
   const ref = useRef<HTMLDivElement>(null);
   useAddHashToUrl(id, ref);
