+++ components/GlobalStyle.ts
