+++ components/LottieAnimation.tsx
