+++ components/VoteTicker.tsx
@@ -20,12 +20,25 @@ import UpRightArrow from "public/assets/up-right-arrow.svg";
 import { useState } from "react";
 import styled, { CSSProperties } from "styled-components";
 import { useInterval } from "usehooks-ts";
-import { formatMillisecondsUntilMidnight } from "./utils";
 
 export function VoteTicker({ isLightTheme = false }) {
   const { data } = useVotingInfo();
   const [timeRemaining, setTimeRemaining] = useState("--:--:--");
   const isActive = !!data && data.activeRequests > 0;
+
+  function getMillisecondsUntilMidnight() {
+    const now = new Date();
+    const midnight = new Date(now.getFullYear(), now.getMonth(), now.getDate() + 1);
+    return midnight.getTime() - now.getTime();
+  }
+
+  function formatMillisecondsUntilMidnight() {
+    const milliseconds = getMillisecondsUntilMidnight();
+    const seconds = Math.floor(milliseconds / 1000);
+    const minutes = Math.floor(seconds / 60);
+    const hours = Math.floor(minutes / 60);
+    return `${hours}:${minutes % 60}:${seconds % 60}`;
+  }
 
   useInterval(() => {
     setTimeRemaining(formatMillisecondsUntilMidnight());
