--- components/index.ts
@@ -1,16 +1,0 @@
-export { AnimatedLink } from "./AnimatedLink/AnimatedLink";
-export { Builder } from "./Builder/Builder";
-export { Divider } from "./Divider/Divider";
-export { Footer } from "./Footer/Footer";
-export { GlobalStyle } from "./GlobalStyle/GlobalStyle";
-export { Header } from "./Header/Header";
-export { Hero } from "./Hero/Hero";
-export { HowItWorks } from "./HowItWorks/HowItWorks";
-export { Layout } from "./Layout/Layout";
-export { Home } from "./pages/Home/Home";
-export { Projects } from "./Projects/Projects";
-export { SupportSection } from "./SupportSection/SupportSection";
-export { default as Tabs } from "./Tabs/Tabs";
-export { VoteParticipation } from "./VoteParticipation/VoteParticipation";
-export { VoteTicker } from "./VoteTicker/VoteTicker";
-export { MailChimpForm } from "./MailChimpForm/MailChimpForm";
