+++ pages/index.tsx
@@ -1,4 +1,4 @@
-import { Home } from "components";
+import { Home } from "components/pages/Home";
 import type { NextPage } from "next";
 
 const HomePage: NextPage = () => {
