+++ stories/Header.stories.tsx
@@ -1,5 +1,5 @@
 import { ComponentMeta, ComponentStory } from "@storybook/react";
-import { Header } from "components";
+import { Header } from "components/Header";
 
 export default {
   title: "Header",
