+++ stories/Hero.stories.tsx
@@ -1,5 +1,5 @@
 import { ComponentMeta, ComponentStory } from "@storybook/react";
-import { Hero } from "components";
+import { Hero } from "components/Hero";
 
 export default {
   title: "Hero",
